Move YamahaYncaSettingEntity to entity.py
